Rename Git to GitDb
