+++ PineBlog.sln
@@ -70,9 +70,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "90 Testing", "90 Testing", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opw.PineBlog.Sample.NuGet", "samples\Opw.PineBlog.Sample.NuGet\Opw.PineBlog.Sample.NuGet.csproj", "{6725D89E-3A46-4756-8F5E-2AFB7E1986FC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opw.PineBlog.Git", "src\Opw.PineBlog.Git\Opw.PineBlog.Git.csproj", "{34D24B82-9800-49B2-B5E5-96B6A317FDEE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opw.PineBlog.GitDb", "src\Opw.PineBlog.GitDb\Opw.PineBlog.GitDb.csproj", "{216B0A5F-1401-49CF-97C8-AA9E1DAC5326}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Opw.PineBlog.Git.Tests", "tests\Opw.PineBlog.Git.Tests\Opw.PineBlog.Git.Tests.csproj", "{17C8F84F-130F-4112-A8EC-A02A9820AA87}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Opw.PineBlog.GitDb.Tests", "tests\Opw.PineBlog.GitDb.Tests\Opw.PineBlog.GitDb.Tests.csproj", "{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -140,14 +140,14 @@ Global
 		{6725D89E-3A46-4756-8F5E-2AFB7E1986FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6725D89E-3A46-4756-8F5E-2AFB7E1986FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6725D89E-3A46-4756-8F5E-2AFB7E1986FC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{34D24B82-9800-49B2-B5E5-96B6A317FDEE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{34D24B82-9800-49B2-B5E5-96B6A317FDEE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{34D24B82-9800-49B2-B5E5-96B6A317FDEE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{34D24B82-9800-49B2-B5E5-96B6A317FDEE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{17C8F84F-130F-4112-A8EC-A02A9820AA87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{17C8F84F-130F-4112-A8EC-A02A9820AA87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{17C8F84F-130F-4112-A8EC-A02A9820AA87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{17C8F84F-130F-4112-A8EC-A02A9820AA87}.Release|Any CPU.Build.0 = Release|Any CPU
+		{216B0A5F-1401-49CF-97C8-AA9E1DAC5326}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{216B0A5F-1401-49CF-97C8-AA9E1DAC5326}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{216B0A5F-1401-49CF-97C8-AA9E1DAC5326}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{216B0A5F-1401-49CF-97C8-AA9E1DAC5326}.Release|Any CPU.Build.0 = Release|Any CPU
+		{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -168,8 +168,8 @@ Global
 		{B8AD9749-650E-406F-BADE-D40ADDA5B4E7} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
 		{17B168D3-4269-440B-AAD8-69A759035B07} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
 		{6725D89E-3A46-4756-8F5E-2AFB7E1986FC} = {5FEBABA0-73A1-464A-8715-19F76F742B22}
-		{34D24B82-9800-49B2-B5E5-96B6A317FDEE} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
-		{17C8F84F-130F-4112-A8EC-A02A9820AA87} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
+		{216B0A5F-1401-49CF-97C8-AA9E1DAC5326} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
+		{732B3BCD-EA7B-4CFC-A9AA-8048189C2AC9} = {F800AFAF-471F-41F4-8F17-0C4FAEC41F6A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DF018025-4A97-49F8-888F-C70856C1862A}
