+++ tests/Opw.PineBlog.GitDb.Tests/Opw.PineBlog.GitDb.Tests.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
     <IsPackable>false</IsPackable>
-    <RootNamespace>Opw.PineBlog.Git</RootNamespace>
+    <RootNamespace>Opw.PineBlog.GitDb</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
